requirements.txt
pandas
streamlit
flask
w1thermsensor
RPi.GPIO
pigpio  # Alternative for PWM (optional)
# For dummy/demo mode (no hardware):
# (no extra requirements) 